requirements.txt
# Canvas Scraper Enhanced - Python Dependencies

# Core dependencies for Canvas API client
aiohttp>=3.8.0
python-dotenv>=0.19.0

# Supabase integration
supabase>=1.0.0

# Database Layer (PostgreSQL/Supabase)
sqlalchemy>=1.4.0
asyncpg>=0.27.0
alembic>=1.8.0
psycopg2-binary>=2.9.0
databases[postgresql]>=0.7.0

# Text Extraction
PyPDF2>=3.0.0
pdfplumber>=0.7.0
python-pptx>=0.6.21
python-docx>=0.8.11
beautifulsoup4>=4.11.0
chardet>=5.0.0
pytesseract>=0.3.10
Pillow>=9.0.0

# Scheduling & Job Queue
APScheduler>=3.9.0
celery>=5.2.0
redis>=4.3.0
kombu>=5.2.0

# Content Processing
python-magic>=0.4.27

# Search & Text Analysis
whoosh>=2.7.4
nltk>=3.7.0
textstat>=0.7.0

# Monitoring & Observability
prometheus-client>=0.14.0
structlog>=22.1.0
python-json-logger>=2.0.0

# HTTP & Networking
httpx>=0.23.0
aiofiles>=0.8.0
tenacity>=8.0.0

# Utilities
click>=8.1.0
pydantic>=1.9.0
python-dateutil>=2.8.0
pytz>=2022.1

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=3.0.0
pytest-mock>=3.7.0
pytest-xdist>=2.5.0
factory-boy>=3.2.0

# Development and quality assurance
black>=22.3.0
flake8>=4.0.0
mypy>=0.950

# Performance & Optimization
uvloop>=0.16.0
orjson>=3.7.0

# Security
cryptography>=37.0.0
passlib>=1.7.4

# Cloud Storage (Optional)
boto3>=1.24.0
azure-storage-blob>=12.12.0

# Monitoring Extensions
psutil>=5.9.0
py-cpuinfo>=8.0.0